# trailmemo/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_trailmemo_pkg_response.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      error_code:
        example: ROUTE_NOT_FOUND
        type: string
      message:
        example: success
        type: string
      request_id:
        type: string
    type: object
...
  internal_handler.ToggleFavoriteRequest:
    properties:
      route_id:
        example: 1
        type: integer
    required:
    - route_id
    type: object
...
host: localhost:8087
info:
  contact:
    email: support@example.com
    name: API Support
  description: 迹忆旅图寻迹旅游后端 API 文档
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://example.com/terms
  title: TrailMemo API
  version: 1.0.0
paths:
...
  /api/v1/favorites/count:
    get:
      consumes:
      - application/json
      description: 获取路线的收藏数量
      parameters:
      - description: 路线ID
        in: query
        name: route_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取收藏数量
      tags:
      - 收藏
  /api/v1/favorites/list:
    get:
      consumes:
      - application/json
      description: 获取用户收藏的所有路线
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户收藏列表
      tags:
      - 收藏
  /api/v1/favorites/status:
    get:
      consumes:
      - application/json
      description: 检查用户是否已收藏指定路线
      parameters:
      - description: 路线ID
        in: query
        name: route_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取收藏状态
      tags:
      - 收藏
  /api/v1/favorites/toggle:
    post:
      consumes:
      - application/json
      description: 对路线进行收藏或取消收藏
      parameters:
      - description: 收藏信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ToggleFavoriteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 收藏/取消收藏
      tags:
      - 收藏
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"